src/deltaData/requirements.txt
pyspark==3.3.2
delta-spark==2.3.0